类型转换

+++ IBApplication/IBApplication/Custom(基础)/IBKit(工具)/MBSocket/Encode/MBSocketPacketEncode.m
@@ -23,8 +23,8 @@
     MBSocketByte *headerBytes = [[MBSocketByte alloc] initWithData:headerData];
     [headerBytes replaceInt16:packet.mark index:0 htons:YES];
     [headerBytes replaceInt16:packet.messageType index:2 htons:YES];
-    [headerBytes replaceInt32:packet.sequence index:4 htonl:YES];
-    [headerBytes replaceInt32:packet.sesssionId index:8 htonl:YES];
+    [headerBytes replaceInt32:(int32_t)packet.sequence index:4 htonl:YES];
+    [headerBytes replaceInt32:(int32_t)packet.sesssionId index:8 htonl:YES];
     [headerBytes replaceInt16:packet.extraHeaderLength index:12 htons:YES];
     [headerBytes replaceInt16:packet.bodyLength index:14 htons:YES];
